Support larger values
Computer now supports longs instead of ints
Tape has expanded memory feature by providing constructor value

+++ src/main/java/net/jimj/adventofcode/year2019/Day5.java
@@ -9,14 +9,14 @@ public class Day5 {
         System.out.println("Computing for Thermal: " + compute(5));
     }
 
-    private static int compute(
+    private static long compute(
             final int controllerId) {
         final Computer computer = Computer.standard();
         computer.compute(
                 Tape.forInput(2019, 5),
                 controllerId);
 
-        int diagnosticCode = 0;
+        long diagnosticCode = 0;
         while (diagnosticCode == 0) {
             diagnosticCode = computer.output();
         }
